Fix ssz types
