+++ packages/beacon-node/test/e2e/doppelganger/doppelganger.test.ts
@@ -1,7 +1,8 @@
 import {describe, afterEach, it, expect} from "vitest";
 import {fromHexString} from "@chainsafe/ssz";
 import {routes} from "@lodestar/api/beacon";
-import {BLSPubkey, Epoch, phase0, Slot, ssz} from "@lodestar/types";
+import {BLSPubkey, Epoch, phase0, Slot} from "@lodestar/types";
+import {ssz as phase0SSZ} from "@lodestar/types/phase0";
 import {ChainConfig} from "@lodestar/config";
 import {SLOTS_PER_EPOCH} from "@lodestar/params";
 import {Validator} from "@lodestar/validator";
@@ -88,7 +89,7 @@ describe.skip("doppelganger / doppelganger test", function () {
 
     const validatorUnderTest = validatorsWithDoppelganger[0];
     const pubKey = validatorUnderTest.validatorStore.votingPubkeys()[0];
-    const beaconBlock = ssz.allForks.phase0.BeaconBlock.defaultValue();
+    const beaconBlock = phase0SSZ.BeaconBlock.defaultValue();
 
     await expect(
       validatorUnderTest.validatorStore.signBlock(fromHexString(pubKey), beaconBlock, bn.chain.clock.currentSlot)
@@ -241,7 +242,7 @@ describe.skip("doppelganger / doppelganger test", function () {
 
     const validatorUnderTest = validatorsWithDoppelganger[0];
     const pubKey = validatorUnderTest.validatorStore.votingPubkeys()[0];
-    const beaconBlock = ssz.allForks.phase0.BeaconBlock.defaultValue();
+    const beaconBlock = phase0SSZ.BeaconBlock.defaultValue();
 
     await expect(
       validatorUnderTest.validatorStore.signBlock(fromHexString(pubKey), beaconBlock, bn.chain.clock.currentSlot)
