+++ packages/types/src/index.ts
@@ -1,6 +1,8 @@
 export * from "./types.js";
 import * as ssz from "./sszTypes.js";
-export {ssz};
+import {sszTypesFor} from "./sszTypes.js";
+import type {SSZTypesFor, SSZInstanceTypesFor} from "./sszTypes.js";
+export {sszTypesFor, SSZTypesFor, SSZInstanceTypesFor, ssz};
 // Typeguards
 export * from "./utils/typeguards.js";
 // String type
